Remove Ruler-loader related code for Grafana rules
