+++ public/app/features/alerting/unified/rule-list/FilterView.tsx
@@ -12,7 +12,7 @@ import { hashRule } from '../utils/rule-id';
 
 import { DataSourceRuleLoader } from './DataSourceRuleLoader';
 import { FilterProgressState, FilterStatus } from './FilterViewStatus';
-import { GrafanaRuleLoader } from './GrafanaRuleLoader';
+import { GrafanaRuleListItem } from './GrafanaRuleListItem';
 import LoadMoreHelper from './LoadMoreHelper';
 import { UnknownRuleListItem } from './components/AlertRuleListItem';
 import { AlertRuleListItemSkeleton } from './components/AlertRuleListItemLoader';
@@ -154,12 +154,18 @@ function FilterViewResults({ filterState }: FilterViewProps) {
           switch (origin) {
             case 'grafana':
               return (
-                <GrafanaRuleLoader
-                  key={key}
-                  ruleIdentifier={{ ruleSourceName: 'grafana', uid: rule.uid }}
+                <GrafanaRuleListItem
+                  rule={rule}
+                  // rulerRule={rulerRule}
                   groupIdentifier={groupIdentifier}
                   namespaceName={ruleWithOrigin.namespaceName}
                 />
+                // <GrafanaRuleLoader
+                //   key={key}
+                //   ruleIdentifier={{ ruleSourceName: 'grafana', uid: rule.uid }}
+                //   groupIdentifier={groupIdentifier}
+                //   namespaceName={ruleWithOrigin.namespaceName}
+                // />
               );
             case 'datasource':
               return <DataSourceRuleLoader key={key} rule={rule} groupIdentifier={groupIdentifier} />;
